Correcciones Conexion a oracle y cambios en libreria de conexion se utiliza net core.

+++ ModelsStore/DTO/TABLES/INVENTARIO.cs
@@ -11,11 +11,11 @@ namespace ModelsStore.DTO.TABLES
         public string CODIGO_PRODUCTO { get; set; }
         public string NOMBRE_PRODUCTO { get; set; }
         public string DESCRIPCION_PRODUCTO { get; set; }
-        public int UNIDADES_EXISTENTES { get; set; }
+        public Int64 UNIDADES_EXISTENTES { get; set; }
         public DateTime FECHA_CARGA { get; set; }
         public DateTime FECHA_INGRESO { get; set; }
         public string UUID_ESTADO { get; set; }
-        public int ACTIVO { get; set; }
+        public Int32 ACTIVO { get; set; }
         public string CODIGO_PROVEEDOR { get; set; }
         public DateTime CADUCIDAD { get; set; }
